altera parametro de converter para unsigned long

+++ includes/ft_printf.h
@@ -27,7 +27,7 @@ typedef struct s_data_format	t_data_format;
 typedef struct s_hex_base_converter {
 	char	*hex_upper_digits;
 	char	*hex_lower_digits;
-	char	*(*convert)(int, char *);
+	char	*(*convert)(unsigned long, char *);
 }	t_hex_base_converter;
 
 t_data_format			*get_char_format(void);
